Use CLOCK_TIMESTAMP instead of NOW
When generating ULIDs inside a transaction, the timestamp of every ULID will be the same because the value of `NOW` is only computed when the transaction begins. This effectively reduces the precision of the ID.

`CLOCK_TIMESTAMP` is resolved when the statement is executed, so each ULID created in a transaction will contain a more accurate timestamp.